Cleanup Scala compilation tests
We only support using Zinc now.
